#107: add theme builder and update catalog and app
