+++ lib/ui/common/layout/layout.dart
@@ -6,4 +6,5 @@ export './noted_image_header.dart';
 export './noted_loading_indicator.dart';
 export './noted_snack_bar.dart';
 export './noted_tab_bar.dart';
+export './noted_theme_builder.dart';
 export './noted_tile.dart';
